computer-vision/car-classification-transfer-learning/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
tqdm>=4.60.0
Pillow>=8.3.0
albumentations>=1.0.0
opencv-python>=4.5.0
mlflow>=1.20.0
PyYAML>=5.4.0
jupyter>=1.0.0
ipywidgets>=7.6.0